requirements.txt
streamlit>=1.32.0
yfinance>=0.2.36
pandas>=2.0.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
google-genai>=0.8.0
google-generativeai>=0.5.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
lxml>=5.0.0
